Apply Arrhenius scaling to 25 degC for temperature-dependent traits

Apply Arrhenius scaling to 25 degC for temperature-dependent traits

arrhenius.scaling.traits(
  data,
  covariates,
  temp.covariates,
  new.temp = 25,
  missing.temp = 25
)

Arguments

data

data frame of data to scale, as returned by query.data()

covariates

data frame of covariates, as returned by query.covariates(). Note that data with no matching covariates will be unchanged.

temp.covariates

names of covariates used to adjust for temperature; if length > 1, order matters (first will be used preferentially)

new.temp

the reference temperature for the scaled traits. Curerntly 25 degC

missing.temp

the temperature assumed for traits with no covariate found. Curerntly 25 degC

Author

Carl Davidson, David LeBauer, Ryan Kelly