List files associated with a container and machine exist in `dbfiles` table

List files associated with a container and machine exist in `dbfiles` table

Usage

dbfile.check(
  type,
  container.id,
  con,
  hostname = PEcAn.remote::fqdn(),
  machine.check = TRUE,
  return.all = FALSE
)

Arguments

type

The type of `dbfile`, as a character. Must be either "Input", "Posterior", or "Model".

container.id

the ID of container type. E.g. if `type` is "Input", this will correspond to the `id` column from the `inputs` table.

con

database connection object

hostname

the name of the host where the file is stored, this will default to the name of the current machine

machine.check

setting to check for file on named host, otherwise will check for any file given container id

return.all

(Logical) If `TRUE`, return all files. If `FALSE`, return only the most recent files.

Value

`data.frame` with the id, filename and pathname of all the files that are associated

Author

Rob Kooper, Alexey Shiklomanov

Examples

if (FALSE) { # \dontrun{
  dbfile.check("Input", 7, dbcon)
} # }