Get trait data from the database.

This will use the following items from settings: - `settings$pfts` - `settings$model$type` - `settings$database$bety` - `settings$database$dbfiles` - `settings$meta.analysis$update`

Usage

get.trait.data(
  pfts,
  modeltype,
  dbfiles,
  database,
  forceupdate,
  write = FALSE,
  trait.names = NULL,
  input_file = NULL
)

Arguments

pfts

the list of pfts to get traits for

modeltype

type of model that is used, this is is used to distinguish between different PFTs with the same name.

dbfiles

location where previous results are found

database

database connection parameters (see `params` argument to [db.query()])

forceupdate

(Logical) If `TRUE`, force a database update whether or not it is needed. If `FALSE`, only update if an update is needed.

write

(Logical) If `TRUE` updated posteriors will be written to BETYdb. Defaults to FALSE.

trait.names

Character vector of trait names to search. If `NULL` (default), use all traits that have a prior for at least one of the `pfts`.

input_file

Path to a CSV file containing prior information. If specified, `database` is not used.

Value

list of PFTs with update posteriorids

Details

If either `input_file` or `settings$pfts$file_path` is provided, it should be a valid path to a CSV (with at least columns `name`, `distn`, `parama`, `paramb`, `n`) and will be used instead of `database` for trait lookup.

Author

David LeBauer, Shawn Serbin, Alexey Shiklomanov