Associate species with a PFT.

adds a list of species to a pft based on USDA Plants acronyms

pft.add.spp(pft, acronym = NULL, ID = NULL, test = TRUE, con = NULL, ...)

Arguments

pft

String name of the PFT in the database

acronym

Specie's Symbols. see http://plants.usda.gov

ID

Species IDs in Bety. You can provide either IDs or Symbols as input, if you provide both ID and acronym, only acronym will be used.

test

Runs the function in test mode. No species are actually added, but checks are run for existing species-PFT pairs, unmatched acronyms, missing species, or duplicate species

con

Database connection object.

...

optional arguements for connecting to database (e.g. password, user name, database)

Value

Function does not return a value but does print out diagnostic statements.

Details

This function is used to add PFT-Species pairs to the database table 'pfts_species'. In the initial implementation the PFT has to be defined already and the species are added based on their USDA Symbol (genus/species acronym). Multiple species can be added at once but only one PFT at a time.

The Symbols object are

Author

Michael C. Dietze, Dongchen Zhang