Function to query data from database for specific species and convert stat to SE

Function to query data from database for specific species and convert stat to SE

Usage

query.data(
  trait,
  spstr,
  con,
  extra.columns = paste("ST_X(ST_CENTROID(sites.geometry)) AS lon,",
    "ST_Y(ST_CENTROID(sites.geometry)) AS lat, "),
  store.unconverted = FALSE,
  ids_are_cultivars = FALSE,
  ...
)

Arguments

trait

trait to query from the database

spstr

IDs of species to query from, as a single comma-separated string

con

database connection

extra.columns

other query terms to pass in. If unspecified, retrieves latitude and longitude

store.unconverted

determines whether or not a copy of the mean and stat fields are returned with _unconverted appended to the column names

ids_are_cultivars

if TRUE, ids is a vector of cultivar IDs, otherwise they are species IDs

...

extra arguments

See also

Author

David LeBauer, Carl Davidson