Query Priors

Query priors associated with a plant functional type and a set of traits.

Usage

query.priors(pft, trstr = NULL, con = NULL, ...)

Arguments

pft

ID number of the PFT in the database

trstr

String of traits to query priors for. If passed as a character vector, it will be concatenated to a single string using [PEcAn.utils::vecpaste()].

con

Database connection object.

...

Optional arguments for connecting to database (e.g. password, user name, database).

Value

`data.frame` of priors for each trait and the given PFT.

Details

If neither `con` nor `...` are provided, this will try to connect to BETY using a `settings` object in the current environment.

Author

David LeBauer, Alexey Shiklomanov

Examples

if (FALSE) { # \dontrun{
  con <- db.open(...)
  query.priors("ebifarm.pavi", c("SLA", "Vcmax", "leaf_width"), con = con)
} # }