Extract trait data from database

Extracts data from database for a given trait and set of species, converts all statistics to summary statistics, and prepares a dataframe for use in meta-analysis. For Vcmax and SLA data, only data collected between April and July are queried, and only data collected from the top of the canopy (canopy height > 0.66). For Vcmax and root_respiration_rate, data are scaled converted from measurement temperature to \(25^oC\) via the arrhenius equation.

Usage

query.trait.data(
  trait,
  spstr,
  con = NULL,
  update.check.only = FALSE,
  ids_are_cultivars = FALSE,
  ...
)

Arguments

trait

is the trait name used in the database, stored in variables.name

spstr

is the species.id integer or string of integers associated with the species

con

database connection object

update.check.only

if TRUE, returns results but does not print summaries

ids_are_cultivars

if TRUE, the IDs in spstr are cultivar IDs, otherwise they are species IDs. Passed on to query.data

...

unused currently

Value

dataframe ready for use in meta-analysis

Author

David LeBauer, Carl Davidson, Shawn Serbin

Examples

if (FALSE) { # \dontrun{
settings <- read.settings()
query.trait.data("Vcmax", "938", con = con)
} # }