Query trait data

Query available trait data associated with a given pft and a list of traits

query.traits(ids, priors, con = NULL, update.check.only = FALSE,
  ids_are_cultivars = FALSE)

Arguments

ids
vector of species or cultivar id's from trait database

priors
vector of parameters for which priors have been specified

con
database connection object

update.check.only
if TRUE, returns results but does not print summaries

ids_are_cultivars
if TRUE, ids is a vector of cultivar IDs, otherwise they are species IDs

Value

list of dataframes, each with data for one trait

See also

Examples

# NOT RUN {
species <- query.pft_species('ebifarm.c4crop')
spstr <- vecpaste(species$id)
trvec <- c('leafN', 'SLA')
trait.data <- query.traits(spstr, trvec)
# }