Get trait data from the database for a single PFT

Get trait data from the database for a single PFT

Usage

get.trait.data.pft(
  pft,
  modeltype,
  dbfiles,
  dbcon,
  trait.names,
  forceupdate = FALSE,
  write = FALSE
)

Arguments

pft

list of settings for the pft whose traits to retrieve. See details

modeltype

type of model that is used, this is used to distinguish between different pfts with the same name.

dbfiles

location where previous results are found

dbcon

database connection

trait.names

list of trait names to retrieve

forceupdate

set this to true to force an update, auto will check to see if an update is needed.

write

(Logical) If `TRUE` updated posteriors will be written to BETYdb. Defaults to FALSE.

Value

updated pft with posteriorid

Details

`pft` should be a list containing at least `name` and `outdir`, and optionally `posteriorid` and `constants`. BEWARE: All existing files in `outir` will be deleted!

Author

David LeBauer, Shawn Serbin, Rob Kooper